+++ core/src/network/rpc.rs
@@ -1,5 +1,6 @@
 use avail_rust::{
 	kate_recovery::matrix::{Dimensions, Position},
+	sp_core::bytes::from_hex,
 	AvailHeader, H256,
 };
 use codec::{Decode, Encode};
@@ -7,7 +8,6 @@ use color_eyre::{eyre::eyre, Result};
 use configuration::RPCConfig;
 use rand::{seq::SliceRandom, thread_rng, Rng};
 use serde::{de, Deserialize, Serialize};
-use sp_core::bytes::from_hex;
 use std::{collections::HashSet, fmt::Display};
 use tokio::{
 	sync::broadcast,