+++ client/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.12.0
 
+- Switch to `avail-rust`
 - Move `crawl` feature to separate bin
 - Binary name changed from `avail-light` to `avail-light-client`
 - Core changes are enumerated in [`core` CHANGELOG](../core/CHANGELOG.md) under the version 1.12.0